@@ -26,72 +26,77 @@
     "test": "jest -i",
     "update-schema": "typescript-json-schema --required --noExtraProps tsconfig.json AppManifest > ./src/app/types/AppManifest.schema.json && typescript-json-schema --required --noExtraProps src/app/types/SchemaObject.ts SchemaObject > ./src/app/types/SchemaObject.schema.json && typescript-json-schema --required --noExtraProps src/app/types/DestinationSchema.ts DestinationSchema > ./src/app/types/DestinationSchema.schema.json",
     "prepare-readme-export": "node scripts/prepare_readme_export.js",
-    "generate-docs": "typedoc --plugin typedoc-plugin-markdown ./src/index.ts && yarn prepare-readme-export"
+    "generate-docs": "typedoc && yarn prepare-readme-export"
   },
   "engines": {
-    "node": ">=18.0"
+    "node": ">=22.0.0"
   },
   "dependencies": {
-    "@zaiusinc/app-forms-schema": "^1.0.1",
-    "ajv": "^8.11.0",
+    "@zaiusinc/app-forms-schema": "^2.0.0",
+    "ajv": "^8.17.1",
     "cron-expression-validator": "^1.0.20",
-    "csv-parser": "^3.0.0",
+    "csv-parser": "^3.2.0",
     "deep-freeze": "^0.0.1",
     "email-validator": "^2.0.4",
-    "glob": "^8.0.3",
+    "glob": "^11.0.1",
     "js-yaml": "^4.1.0",
     "jsonpath": "^1.1.1",
-    "node-fetch": "^2.6.7",
+    "node-fetch": "^2.7.0",
     "object-hash": "^3.0.0",
     "remark": "^13.0.0",
-    "remark-validate-links": "^9.0.1"
+    "remark-validate-links": "^10.0.4"
   },
   "devDependencies": {
-    "@types/deep-freeze": "^0.1.2",
-    "@types/glob": "^8.0.0",
-    "@types/jest": "^29.5.6",
-    "@types/js-yaml": "^4.0.5",
-    "@types/jsonpath": "^0.2.3",
-    "@types/lodash": "^4.14.200",
-    "@types/mock-fs": "^4.13.1",
-    "@types/node": "^18.7.13",
-    "@types/node-fetch": "^2.6.2",
-    "@types/object-hash": "^2.2.1",
-    "@types/url-regex-safe": "^1.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.35.1",
-    "@typescript-eslint/eslint-plugin-tslint": "^5.35.1",
-    "@typescript-eslint/parser": "^5.0.0",
-    "@zaiusinc/eslint-config-presets": "^1.0.0",
-    "@zaiusinc/node-sdk": "^1.0.0",
-    "eslint": "^8.52.0",
-    "eslint-config-standard-with-typescript": "^22.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.4.3",
-    "eslint-plugin-jsdoc": "^39.3.6",
-    "eslint-plugin-n": "^15.0.0",
+    "@eslint/compat": "^1.2.7",
+    "@stylistic/eslint-plugin": "^4.2.0",
+    "@types/deep-freeze": "^0.1.5",
+    "@types/glob": "^8.1.0",
+    "@types/jest": "^29.5.14",
+    "@types/js-yaml": "^4.0.9",
+    "@types/jsonpath": "^0.2.4",
+    "@types/lodash": "^4.17.16",
+    "@types/mock-fs": "^4.13.4",
+    "@types/node": "^22.13.10",
+    "@types/node-fetch": "^2.6.12",
+    "@types/object-hash": "^3.0.6",
+    "@types/url-regex-safe": "^1.0.2",
+    "@typescript-eslint/eslint-plugin": "^8.27.0",
+    "@typescript-eslint/eslint-plugin-tslint": "^7.0.2",
+    "@typescript-eslint/parser": "^8.27.0",
+    "@zaiusinc/eslint-config-presets": "^2.0.0",
+    "@zaiusinc/node-sdk": "^2.0.0",
+    "eslint": "^9.22.0",
+    "eslint-config-love": "^119.0.0",
+    "eslint-config-standard-with-typescript": "^43.0.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-jsdoc": "^50.6.8",
+    "eslint-plugin-n": "^17.16.2",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.31.1",
+    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-react": "^7.37.4",
     "jest": "^29.7.0",
+    "jest-junit": "^16.0.0",
     "lodash": "^4.17.21",
-    "mock-fs": "^5.1.4",
-    "nock": "^13.2.9",
-    "prettier": "^2.8.7",
+    "mock-fs": "^5.5.0",
+    "nock": "^14.0.1",
+    "prettier": "^3.5.3",
     "slugify": "^1.6.6",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
+    "ts-jest": "^29.2.6",
+    "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
-    "tslint": "^5.14.0",
-    "typedoc": "^0.24.0",
-    "typedoc-plugin-markdown": "=4.0.0-next.4",
-    "typescript": "^4.7.4",
-    "typescript-json-schema": "^0.54.0",
-    "jest-junit": "^16.0.0"
+    "tslint": "^6.1.3",
+    "typedoc": "^0.28.4",
+    "typedoc-plugin-frontmatter": "^1.3.0",
+    "typedoc-plugin-markdown": "^4.6.3",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.27.0",
+    "typescript-json-schema": "^0.65.1"
   },
   "peerDependencies": {
-    "@zaiusinc/node-sdk": "^1.0.0"
+    "@zaiusinc/node-sdk": "^2.0.0"
   },
   "resolutions": {
-    "@babel/core": "^7.23.2"
+    "@babel/core": "^7.26.10"
   }
 }
