@@ -1,9 +1,9 @@
 import Ajv from 'ajv';
-import * as deepFreeze from 'deep-freeze';
-import * as fs from 'fs';
+import deepFreeze from 'deep-freeze';
+import fs from 'fs';
 import 'jest';
 import * as jsYaml from 'js-yaml';
-import * as mockFs from 'mock-fs';
+import mockFs from 'mock-fs';
 import {Runtime} from './Runtime';
 import {AppManifest} from './types';
 import {SchemaObject} from './types/SchemaObject';
@@ -262,8 +262,8 @@ describe('Runtime', () => {
 
       const result = runtime.getSchemaObjects();
       expect(readFileSyncFn.mock.calls).toEqual([
+        ['/tmp/foo/schema/my_app_coupons.yaml', 'utf8'],
         ['/tmp/foo/schema/events.yml', 'utf8'],
-        ['/tmp/foo/schema/my_app_coupons.yaml', 'utf8']
       ]);
       expect(result).toEqual(schemaObjects);
 
