@@ -1,4 +1,4 @@
-import * as chalk from 'chalk';
+import chalk from 'chalk';
 import * as path from 'path';
 import {Runtime} from '../Runtime';
 import {validateApp} from './validateApp';
