@@ -1,5 +1,5 @@
-import * as deepFreeze from 'deep-freeze';
-import * as mockFs from 'mock-fs';
+import deepFreeze from 'deep-freeze';
+import mockFs from 'mock-fs';
 import {Runtime} from '../../Runtime';
 import {AppManifest} from '../../types';
 import {validateAssets} from '../validateAssets';
