@@ -1,4 +1,4 @@
-import * as deepFreeze from 'deep-freeze';
+import deepFreeze from 'deep-freeze';
 import 'jest';
 import {Runtime} from '../../Runtime';
 import {AppManifest} from '../../types';
