@@ -1,7 +1,7 @@
 /* eslint max-classes-per-file: "off" */
 import 'jest';
 import {CsvRowProcessor, CsvStream} from '../CsvStream';
-import * as nock from 'nock';
+import nock from 'nock';
 import {Stream} from 'stream';
 import {Options} from 'csv-parser';
 import * as ObjectHash from 'object-hash';
