@@ -1,7 +1,7 @@
 /* eslint max-classes-per-file: "off" */
 import 'jest';
 import {JsonLinesStream} from '../JsonLinesStream';
-import * as nock from 'nock';
+import nock from 'nock';
 import * as zlib from 'zlib';
 import {Stream} from 'stream';
 import * as ObjectHash from 'object-hash';
